+++ src/Detach/Extensions/VectorExtensions.cs
@@ -7,17 +7,17 @@ public static class VectorExtensions
 {
 	public static Vector2 Round(this Vector2 vector, int digits)
 	{
-		return new(MathF.Round(vector.X, digits), MathF.Round(vector.Y, digits));
+		return new Vector2(MathF.Round(vector.X, digits), MathF.Round(vector.Y, digits));
 	}
 
 	public static Vector3 Round(this Vector3 vector, int digits)
 	{
-		return new(MathF.Round(vector.X, digits), MathF.Round(vector.Y, digits), MathF.Round(vector.Z, digits));
+		return new Vector3(MathF.Round(vector.X, digits), MathF.Round(vector.Y, digits), MathF.Round(vector.Z, digits));
 	}
 
 	public static Vector4 Round(this Vector4 vector, int digits)
 	{
-		return new(MathF.Round(vector.X, digits), MathF.Round(vector.Y, digits), MathF.Round(vector.Z, digits), MathF.Round(vector.W, digits));
+		return new Vector4(MathF.Round(vector.X, digits), MathF.Round(vector.Y, digits), MathF.Round(vector.Z, digits), MathF.Round(vector.W, digits));
 	}
 
 	public static bool IsReal(this Vector2 vector)