+++ src/Detach/Parsers/Model/ObjFormat/ObjParser.cs
@@ -16,16 +16,15 @@ public static class ObjParser
 		string currentObject = string.Empty;
 		string currentGroup = string.Empty;
 		string currentMaterial = string.Empty;
-		for (int i = 0; i < lines.Length; i++)
+		foreach (string line in lines)
 		{
-			string line = lines[i];
 			string[] values = line.Split(' ');
 
 			switch (values[0])
 			{
-				case "v": context.Positions.Add(new(ParseVertexFloat(values[1]), ParseVertexFloat(values[2]), ParseVertexFloat(values[3]))); break;
-				case "vt": context.Textures.Add(new(ParseVertexFloat(values[1]), ParseVertexFloat(values[2]))); break;
-				case "vn": context.Normals.Add(new(ParseVertexFloat(values[1]), ParseVertexFloat(values[2]), ParseVertexFloat(values[3]))); break;
+				case "v": context.Positions.Add(new Vector3(ParseVertexFloat(values[1]), ParseVertexFloat(values[2]), ParseVertexFloat(values[3]))); break;
+				case "vt": context.Textures.Add(new Vector2(ParseVertexFloat(values[1]), ParseVertexFloat(values[2]))); break;
+				case "vn": context.Normals.Add(new Vector3(ParseVertexFloat(values[1]), ParseVertexFloat(values[2]), ParseVertexFloat(values[3]))); break;
 				case "o": currentObject = values[1].Trim(); break;
 				case "g": currentGroup = values[1].Trim(); break;
 				case "usemtl": currentMaterial = values[1].Trim(); break;
@@ -38,7 +37,7 @@ public static class ObjParser
 					for (int j = 0; j < rawIndices.Length; j++)
 					{
 						string[] indexEntries = rawIndices[j].Split('/');
-						faces.Add(new(ushort.Parse(indexEntries[0]), ushort.TryParse(indexEntries[1], out ushort texture) ? texture : (ushort)0, ushort.Parse(indexEntries[2])));
+						faces.Add(new Face(ushort.Parse(indexEntries[0], CultureInfo.InvariantCulture), ushort.TryParse(indexEntries[1], out ushort texture) ? texture : (ushort)0, ushort.Parse(indexEntries[2], CultureInfo.InvariantCulture)));
 
 						if (j >= 3)
 						{
@@ -52,7 +51,7 @@ public static class ObjParser
 						MeshBuildingContext? mesh = context.Meshes.Find(m => m.ObjectName == currentObject);
 						if (mesh == null)
 						{
-							mesh = new()
+							mesh = new MeshBuildingContext
 							{
 								GroupName = currentGroup,
 								MaterialName = currentMaterial,
@@ -69,12 +68,12 @@ public static class ObjParser
 		}
 
 		List<MeshData> meshes = context.Meshes.ConvertAll(mbc => new MeshData(mbc.ObjectName, mbc.GroupName, mbc.MaterialName, mbc.Faces));
-		return new(context.Positions, context.Textures, context.Normals, meshes);
+		return new ModelData(context.Positions, context.Textures, context.Normals, meshes);
 	}
 
 	private static float ParseVertexFloat(string value)
 	{
-		return (float)double.Parse(value, NumberStyles.Float);
+		return (float)double.Parse(value, NumberStyles.Float, CultureInfo.InvariantCulture);
 	}
 
 	private sealed class ModelBuildingContext