+++ src/Detach/Parsers/Texture/TextureParseException.cs
@@ -1,8 +1,5 @@
-using System.Runtime.Serialization;
-
 namespace Detach.Parsers.Texture;
 
-[Serializable]
 public class TextureParseException : Exception
 {
 	public TextureParseException()
@@ -16,11 +13,6 @@ public class TextureParseException : Exception
 
 	public TextureParseException(string? message, Exception? innerException)
 		: base(message, innerException)
-	{
-	}
-
-	protected TextureParseException(SerializationInfo info, StreamingContext context)
-		: base(info, context)
 	{
 	}
 }