+++ src/Detach/Utils/VectorUtils.cs
@@ -20,17 +20,17 @@ public static class VectorUtils
 
 	public static Vector2 Clamp(Vector2 vector, float min, float max)
 	{
-		return new(Math.Clamp(vector.X, min, max), Math.Clamp(vector.Y, min, max));
+		return new Vector2(Math.Clamp(vector.X, min, max), Math.Clamp(vector.Y, min, max));
 	}
 
 	public static Vector3 Clamp(Vector3 vector, float min, float max)
 	{
-		return new(Math.Clamp(vector.X, min, max), Math.Clamp(vector.Y, min, max), Math.Clamp(vector.Z, min, max));
+		return new Vector3(Math.Clamp(vector.X, min, max), Math.Clamp(vector.Y, min, max), Math.Clamp(vector.Z, min, max));
 	}
 
 	public static Vector4 Clamp(Vector4 vector, float min, float max)
 	{
-		return new(Math.Clamp(vector.X, min, max), Math.Clamp(vector.Y, min, max), Math.Clamp(vector.Z, min, max), Math.Clamp(vector.W, min, max));
+		return new Vector4(Math.Clamp(vector.X, min, max), Math.Clamp(vector.Y, min, max), Math.Clamp(vector.Z, min, max), Math.Clamp(vector.W, min, max));
 	}
 
 	public static Vector3 Reflect(Vector3 direction, Vector3 normal)
@@ -48,7 +48,7 @@ public static class VectorUtils
 		float magnitude = MathF.Sqrt(sqrMagnitude);
 		float normalizedX = vector.X / magnitude;
 		float normalizedY = vector.Y / magnitude;
-		return new(normalizedX * maxLength, normalizedY * maxLength);
+		return new Vector2(normalizedX * maxLength, normalizedY * maxLength);
 	}
 
 	public static Vector3 ClampMagnitude(Vector3 vector, float maxLength)
@@ -61,6 +61,6 @@ public static class VectorUtils
 		float normalizedX = vector.X / magnitude;
 		float normalizedY = vector.Y / magnitude;
 		float normalizedZ = vector.Z / magnitude;
-		return new(normalizedX * maxLength, normalizedY * maxLength, normalizedZ * maxLength);
+		return new Vector3(normalizedX * maxLength, normalizedY * maxLength, normalizedZ * maxLength);
 	}
 }